db/requirements.txt
 mysql-connector-python==8.3.0
 flask==3.0.1
 flask-restful==0.3.10
